ui | create gallery module; move photostore to gallery module;
