+++ src/app/modules/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SignInComponent } from './sign-in/sign-in.component';
-import { SignUpComponent } from './sign-up/sign-up.component';
-import { AuthModalComponent } from './auth-modal/auth-modal.component';
+import { SignInComponent } from './components/sign-in/sign-in.component';
+import { SignUpComponent } from './components/sign-up/sign-up.component';
+import { AuthModalComponent } from './components/auth-modal/auth-modal.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ShowPasswordDirective } from 'src/app/directives/show-password.directive';
