+++ src/app/modules/gallery/components/show-image/show-image.component.css
