+++ src/app/modules/gallery/store/photo.actions.ts
