+++ src/app/modules/gallery/store/photo.state.ts
