+++ src/app/store/reducers/app.reducer.ts
@@ -2,7 +2,7 @@ import { routerReducer } from "@ngrx/router-store";
 import { ActionReducerMap } from "@ngrx/store";
 import { IAppState } from "../states/app.state";
 import { authReducer } from "../../modules/auth/store/auth.reducer";
-import { photosReducer } from "./photos.reducer";
+import { photosReducer } from "../../modules/gallery/store/photos.reducer";
 import { userReducer } from "./user.reducer";
 
 export const appReducer: ActionReducerMap<IAppState, any> = {
