+++ src/app/store/states/app.state.ts
@@ -1,6 +1,6 @@
 import { RouterReducerState } from '@ngrx/router-store';
 import { initialUserState, IUserState } from './user.state';
-import { initialPhotoState, IPhotoState } from './photo.state';
+import { initialPhotoState, IPhotoState } from '../../modules/gallery/store/photo.state';
 import { IAuthState, initialAuthState } from '../../modules/auth/store/auth.state';
 
 export interface IAppState {
